✨ feat(kubernetes): add new HelmRelease for homepage app in helmrelease.yaml
This commit introduces a new HelmRelease for the homepage application. It includes configuration for the application's controllers, service, ingress, persistence, and service account. The HelmRelease is set to use the app-template chart version 2.5.0 from the HelmRepository named bjw-s in the flux-system namespace. The application's image is set to ghcr.io/gethomepage/homepage:v0.8.7. The HelmRelease also includes configuration for environment variables, resource requests and limits, ingress settings, and persistence settings.