✨ feat(kubernetes): add new storageclass.yaml for NFS CSI driver
This commit introduces a new storage class for the NFS CSI driver. This will allow Kubernetes to provision persistent volumes using the NFS server. The storage class is configured to use the NFS server at 192.168.86.44 and the share /mnt/tank/db. The reclaim policy is set to Delete and the volume binding mode is set to Immediate.